requirements.txt
certifi==2022.12.7
chardet==3.0.4
click==8.1.3
Flask==2.2.3
idna==2.10
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
requests
urllib3
Werkzeug==2.2.3
zipp==3.15.0
python-dotenv==0.21.1
h3==3.7.6
elasticsearch==8.6.0
python-telegram-bot==13.15